+++ lib/presentation/screens/splash_screen.dart
@@ -1,0 +1,35 @@
+import 'package:fintek/core/constants/images.dart';
+import 'package:fintek/core/constants/strings.dart';
+import 'package:flutter/material.dart';
+
+class SplashScreen extends StatefulWidget {
+  const SplashScreen({Key? key}) : super(key: key);
+
+  @override
+  State<SplashScreen> createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      child: Scaffold(
+        backgroundColor: Colors.black,
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Image.asset(AppImages.logo),
+            const CircularProgressIndicator(),
+            const SizedBox(height: 5),
+            const Text(
+              AppStrings.appVersion,
+              style: TextStyle(
+                color: Colors.white
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}